evse.h, evse.cpp, OneWire.cpp: expand RFIDlist from 20 cards to 100

+++ SmartEVSE-3/include/evse.h
@@ -459,7 +459,7 @@ extern uint32_t ScrollTimer;
 extern uint8_t ChargeDelay;                                                     // Delays charging in seconds.
 extern uint8_t TestState;
 extern uint8_t Access_bit;
-extern uint8_t CardOffset;
+extern uint16_t CardOffset;
 
 extern uint8_t GridActive;                                                      // When the CT's are used on Sensorbox2, it enables the GRID menu option.
 extern uint8_t CalActive;                                                       // When the CT's are used on Sensorbox(1.5 or 2), it enables the CAL menu option.
